+++ src/options.go
@@ -57,6 +57,8 @@ const usage = `usage: fzf [options]
   Layout
     --height=[~]HEIGHT[%]  Display fzf window below the cursor with the given
                            height instead of using fullscreen.
+                           A negative value is calcalated as the terminal height
+                           minus the given value.
                            If prefixed with '~', fzf will determine the height
                            according to the input size.
     --min-height=HEIGHT    Minimum height when --height is given in percent
@@ -157,6 +159,7 @@ type heightSpec struct {
 	size    float64
 	percent bool
 	auto    bool
+	inverse bool
 }
 
 type sizeSpec struct {
@@ -1384,6 +1387,13 @@ func parseHeight(str string) heightSpec {
 	heightSpec := heightSpec{}
 	if strings.HasPrefix(str, "~") {
 		heightSpec.auto = true
+		str = str[1:]
+	}
+	if strings.HasPrefix(str, "-") {
+		if heightSpec.auto {
+			errorExit("negative(-) height is not compatible with adaptive(~) height")
+		}
+		heightSpec.inverse = true
 		str = str[1:]
 	}